@@ -124,7 +124,7 @@ namespace MultiplayerPrairieKing.Entities
 							PK_SpikeyNewTarget message = new();
 							message.target = targetPosition;
 							message.id = id;
-							gameInstance.modInstance.Helper.Multiplayer.SendMessage(message, "PK_SpikeyNewTarget");
+							gameInstance.modInstance.SyncMessage(message);
 						}
 						break;
 					}
@@ -163,7 +163,7 @@ namespace MultiplayerPrairieKing.Entities
                     which = (int)which,
                     position = position
                 };
-                game.modInstance.Helper.Multiplayer.SendMessage(message, "PK_EnemySpawn");
+                game.modInstance.SyncMessage(message);
 			}
 		}
 
@@ -320,7 +320,7 @@ namespace MultiplayerPrairieKing.Entities
 									PK_SpikeyNewTarget message = new();
 									message.id = id;
 									message.target = targetPosition;
-									gameInstance.modInstance.Helper.Multiplayer.SendMessage(message, "PK_SpikeyNewTarget");
+									gameInstance.modInstance.SyncMessage(message);
 								} 
 							}
 						}
@@ -418,7 +418,7 @@ namespace MultiplayerPrairieKing.Entities
                                     {
                                         id = gameInstance.monsters[i].id
                                     };
-                                    gameInstance.modInstance.Helper.Multiplayer.SendMessage(message, "PK_EnemyKilled");
+                                    gameInstance.modInstance.SyncMessage(message);
 
 									gameInstance.AddGuts(gameInstance.monsters[i].position.Location, gameInstance.monsters[i].type);
 									Game1.playSound("Cowboy_monsterDie");
@@ -426,9 +426,11 @@ namespace MultiplayerPrairieKing.Entities
 								}
 							}
 						}
-						if (gameInstance.map.IsCollidingWithMapForMonsters(attemptedPosition) || gameInstance.map.IsCollidingWithMonster(attemptedPosition, this) || gameInstance.player.deathTimer > 0f)
+						if (gameInstance.map.IsCollidingWithMapForMonsters(attemptedPosition) || gameInstance.map.IsCollidingWithMonster(attemptedPosition, this, gameInstance) || gameInstance.player.deathTimer > 0f)
 						{
-							break;
+							gameInstance.map.IsCollidingWithMonster(attemptedPosition, this, gameInstance);
+                            break;
+
 						}
 						ticksSinceLastMovement = 0;
 						position = attemptedPosition;
@@ -454,7 +456,7 @@ namespace MultiplayerPrairieKing.Entities
 								SpikeyStartTransform();
 								PK_SpikeyTransform message = new();
 								message.id = id;
-								gameInstance.modInstance.Helper.Multiplayer.SendMessage(message, "PK_SpikeyTransform");
+								gameInstance.modInstance.SyncMessage(message);
 							}
 						}
 						break;
@@ -493,7 +495,7 @@ namespace MultiplayerPrairieKing.Entities
 						{
 							acceleration.Y -= 0.1f * accelerationMultiplyer;
 						}
-						if (!gameInstance.map.IsCollidingWithMonster(new Rectangle(position.X + (int)Math.Ceiling(acceleration.X), position.Y + (int)Math.Ceiling(acceleration.Y), TileSize, TileSize), this) && gameInstance.player.deathTimer <= 0f)
+						if (!gameInstance.map.IsCollidingWithMonster(new Rectangle(position.X + (int)Math.Ceiling(acceleration.X), position.Y + (int)Math.Ceiling(acceleration.Y), TileSize, TileSize), this, gameInstance) && gameInstance.player.deathTimer <= 0f)
 						{
 							ticksSinceLastMovement = 0;
 							position.X += (int)Math.Ceiling(acceleration.X);
